scrape_event 함수 단순화 및 DB 접근 유틸 별도 모듈로 분리
